Armor and SuperTools
Fixed an Armor bug which lead to crashed on the server as the method was clientsided. Added the a 3x3 function for the Super Pickaxe, Super Shovel(left and right click) and Hoe. The Super Axe can chop down a whole tree as long as it is the same wood.